refactor: extract telemetry rate logic into TelemetryController
Resolves #25.

Moves readOnlyMessageRate, messageRates, availableDataList,
setMessageRate(), processDataStreamRequest(), processMessageInterval(),
and sendMessages() from MAVLinkController into a new TelemetryController
class implementing IController.

MAVLinkController now delegates rate locking/unlocking via
lockRates()/unlockRates() and message sending via
sendMessages(controllers, aircraft), keeping its own role as a
routing coordinator rather than a telemetry manager.

Additional cleanup in MAVLinkController:
- client changed from nullable to lateinit var
- Added typed computed properties for navigationController,
  parameterController, and telemetryController, replacing
  inline filterIsInstance calls at usage sites
- availableDataList made private in TelemetryController
- Removed redundant empty-list guard before sendMessages loop
